\name{dpla_basic}
\alias{dpla_basic}
\title{Search metadata from the Digital Public Library of America (DPLA).}
\usage{
  dpla_basic(q = NULL)
}
\arguments{
  \item{q}{Query terms.}
}
\value{
  A list for now...
}
\description{
  Search metadata from the Digital Public Library of
  America (DPLA).
}
\examples{
\dontrun{
# Basic search, "fruit" in any fields
dpla_basic("fruit")
}
}

