% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eu_providers}
\alias{eu_providers}
\title{Europeana data providers.}
\usage{
eu_providers(providerid = NULL, datasetid = NULL, datasets = FALSE,
  offset = NULL, limit = 10, country_code = NULL, key = NULL,
  callopts = list())
}
\arguments{
\item{providerid}{Provider ID}

\item{datasetid}{Dataset ID}

\item{datasets}{(logical) Whether to return datasets or not with provider information. Ignored
when providerid is NULL (the default for that parameter).}

\item{offset}{(numeric) Query for paging needs - offset to start with}

\item{limit}{(numeric) query    for paging needs - size of the result set to fetch}

\item{country_code}{(character) query    two-letters ISO 3166-1 country code. (TBD - is it so?)}

\item{callopts}{Curl options passed on to httr::GET}
}
\description{
Europeana data providers.
}
\examples{
\dontrun{
out <- eu_providers(limit=3)
out$meta
out$items
eu_providers(providerid=20239)
eu_providers(providerid=20239, datasets=TRUE)
eu_providers(datasetid=2023901)
identical(eu_providers(providerid=20239, datasets=TRUE)$items,
   eu_providers(datasetid=2023901)$items)

eu_providers(country_code='US')
}
}

