% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eu_search}
\alias{eu_search}
\title{Europeana search.}
\usage{
eu_search(query, key = NULL, callopts = list())
}
\arguments{
\item{query}{}

\item{callopts}{Curl options passed on to httr::GET}
}
\description{
Europeana search.
}
\examples{
\dontrun{
eu_search(query='Mona Lisa')
eu_search(query='who:"Leonardo da Vinci"')
eu_search(query='mona AND lisa')
out <- eu_search(query='Lasioglossum')
out$itemsCount
out$items[[6]]
library("httr")
GET(out$items[[6]]$link)
browseURL(out$items[[6]]$guid)
}
}

