% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dpla_plot}
\alias{dpla_plot}
\title{Search metadata from the Digital Public Library of America (DPLA).}
\usage{
dpla_plot(input, plottype = "subjectsum")
}
\arguments{
\item{input}{A data.frame from calling a variety of dpla_* functions.}

\item{plottype}{One of a number of types. These are for convenience. You
can of course create your own plots with more flexibility.}
}
\value{
A ggplot2 object, a plot that is.
}
\description{
Search metadata from the Digital Public Library of America (DPLA).
}
\examples{
\dontrun{
# Plot results, summarising subjects by number of subjects listed
out <- dpla_basic(q="ecology", fields=c("publisher","subject"), limit=90)
dpla_plot(input=out)

# Another example
out <- dpla_basic(q="science", date.before=1900, limit=200)
dpla_plot(input=out, plottype="subjectsbydate")
}
}

